JS. getSnapshot. Added debug flag.
https://trello.com/c/RYIRZksG/13869-glitches-in-reader-on-ipad-pro-after-opening-side-panel

+++ platforms/js/DisplayObjectHelper.hx
@@ -632,7 +632,9 @@ class DisplayObjectHelper {
 			clip.mask = null;
 			untyped clip.maskContainer = null;
 
-			deleteNativeWidget(clip);
+			if (Util.getParameter("debug_snapshot") != '1') {
+				deleteNativeWidget(clip);
+			}
 
 			invalidateTransform(clip, 'removeScrollRect');
 		}
